xephyr: propagate captured raw pointer input
Mirror explicit nested pointer grabs to the focused host window and forward host XI2 RawMotion while captured, so first-person controls continue beyond host window and monitor edges.

Release capture on focus loss or SIGUSR2, retain a recentering fallback for hosts without XI2, and extend the nested integration test with edge motion, focus release, reacquisition, and emergency release coverage.

+++ hw/kdrive/src/kdrive.h
@@ -402,7 +402,8 @@ KdEnqueuePointerEvent(KdPointerInfo * pi, unsigned long flags, int rx, int ry,
 void
 KdEnqueuePointerMotionWithRawDeltas(KdPointerInfo * pi, unsigned long flags,
                                     int rx, int ry, int rz,
-                                    int raw_rx, int raw_ry, int raw_rz);
+                                    double raw_rx, double raw_ry,
+                                    double raw_rz);
 
 void
  KdSetPointerMatrix(KdPointerMatrix *pointer);
